About to refactor away from this design in favour of something a bit less unwieldly.
This works okay so far, but enabling/disabling callbacks on the fly, as well as the central state manager being involved for every transaction is bloating code substantially now that I'm implementing gym_timer program logic.

I would prefer the active program to provide a handle to their callbacks when they 'activate'

Additionally, transition menu task to be a concrete state manager, rather than having a state manager that provides little other than to swap the state, and enable the callbacks, firing up a 'lower' level state machine once a program is selected. Feels way cleaner to just treat the menu as the entry point to the state machine, and have the "Idle" state run as a program, rather than an external set of logic to display time of day.